update class FilmService add validateFilm
