+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/InMemoryFilmStorage.java
@@ -30,7 +30,6 @@ public class InMemoryFilmStorage implements FilmStorage {
 
     @Override
     public Film create(Film film) {
-        validateFilm(film);
         film.setId(getNextId());
         films.put(film.getId(), film);
         return film;
@@ -45,25 +44,10 @@ public class InMemoryFilmStorage implements FilmStorage {
         return ++currentMaxId;
     }
 
-    private void validateFilm(Film film) {
-        if (film.getName() == null || film.getName().isBlank()) {
-            throw new ConditionsNotMetException("Название не может быть пустым");
-        }
-        if (film.getDescription().length() > 200) {
-            throw new ConditionsNotMetException("Описание не может быть длинее 200 символов");
-        }
-        if (film.getDuration() <= 0) {
-            throw new ConditionsNotMetException("Продолжительность фильма должна быть положительным числом.");
-        }
-        if (film.getReleaseDate().isBefore(LocalDate.of(1895,12,28))) {
-            throw new ConditionsNotMetException("Дата релиза должна быть не раньше 28 декабря 1895 года");
-        }
-    }
 
     @Override
     public Film update(Film newFilm) {
             Film oldFilm = films.get(newFilm.getId());
-            validateFilm(newFilm);
             oldFilm.setName(newFilm.getName());
             oldFilm.setDescription(newFilm.getDescription());
             oldFilm.setDuration(newFilm.getDuration());
